Tool/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
alembic>=1.13.1
tenacity>=8.2.3
openai>=1.0.0
google-generativeai>=0.3.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-multipart>=0.0.9
